style: Fix code styling and formatting issues
Using the latest version of the PHP CS configuration, tests for Belgium,
Georgia, the Netherlands, Sweden, and the United Kingdom are updated to
use the most recent code style and formatting rules.

For the most part this updates the header comment and the strict type
declaration. It should not have any affect on the working of the tests
nor the main code.

Signed-off-by: Sacha Telgenhof <me@sachatelgenhof.com>