style: fix code styling and formatting issues (#338)
Using the latest version of the PHP CS configuration, update all source code to use the most recent code style and formatting rules. For the most part this updates the header comment and the strict type declaration. It should not have any affect on the working of the tests nor the main code.

Signed-off-by: Sacha Telgenhof <me@sachatelgenhof.com>