feat(mexico): add Transmission of Federal Executive Power Holiday (#361)
In Mexico presidential elections are held every six years. The new president is inaugurated on 1 October which is a public holiday.

Prior to 2024, this happened on 1 December.

Source:
https://www.gob.mx/profedet/articulos/sabes-cuales-son-los-dias-de-descanso-obligatorios-163134
https://mexico.justia.com/federales/leyes/ley-federal-del-trabajo/titulo-tercero/capitulo-iii/#articulo-74
https://sil.gobernacion.gob.mx/Archivos/Documentos/2022/03/asun_4332475_20220315_1646845246.pdf

I am not sure when this holiday was officially established, so I arbitrarily chose 1934 where the six-year presidential term was introduced. The holiday is defined by the Federal Labour Act (_LeyFederal del Trabajo_). The current law is from 1970, and its predecessor is from 1931.

Source:
https://en.wikipedia.org/wiki/List_of_heads_of_state_of_Mexico#Modern_Mexico_(1934%E2%80%93present)
https://es.wikipedia.org/wiki/Ley_Federal_del_Trabajo_(M%C3%A9xico)

---------

Co-authored-by: Sacha Telgenhof <me@sachatelgenhof.com>