feat(argentina): Movable holidays (#367)
In Argentina, four of the official holidays are movable (_feriados nacionales trasladables_):
- Anniversary of the Passing of General Martín Miguel de Güemes (_Paso ala Inmortalidad del General Don Martín Miguel de Güemes_)
- General José de San Martín Memorial Day (_Paso a la Inmortalidad del General Don José de San Martín_)
- Day of Respect for Cultural Diversity (_Día del Respeto a la Diversidad Cultural_)
- National Sovereignty Day (_Día de la Soberanía Nacional_)

According to the Act on Holidays and Long Weekends (_Ley de establecimiento de feriados y fines de semanas largos_), if a movable holiday falls on a Tuesday or Wednesday, it is moved to the previous Monday; if it falls on a Thursday or Friday, it is moved to the following Monday.

According to Executive Order 789/2021 (_Decreto 789/2021_), if the adjusted holiday coincides with an immovable official holiday, the movable holiday should be observed on its original date, i.e. not moved to a Monday. This exception was applied in 2022, where the adjusted Anniversary of the Passing of General Martín Miguel de Güemes (originally on 17 June) would otherwise be moved to coincide with General Manuel Belgrano Memorial Day on 20 June.

The current rules were adopted in 2017 (the exception for coinciding holidays introduced in 2021 had not been relevant in the years 2017–2021). Previously there were different rules for movable holidays. This PR adds support for the current rules only.

Sources:
https://www.argentina.gob.ar/normativa/nacional/ley-27399-281835/texto
https://www.argentina.gob.ar/normativa/nacional/decreto-789-2021-356678/texto
https://www.rentastucuman.gob.ar/nomina/rentastuc2/nwx1ut2pa3lo/novedades/RG39_22.pdf