fix(ukraine): Ukraine 2021-2023 changes (#369)
This PR implements several changes to public holidays in Ukraine in 2021-2023.

## 2021
In 2021, Day of Defenders of Ukraine was renamed to refer to defenders in a gender-neutral way.

Source:
https://uk.wikipedia.org/wiki/%D0%94%D0%B5%D0%BD%D1%8C_%D0%B7%D0%B0%D1%85%D0%B8%D1%81%D0%BD%D0%B8%D0%BA%D1%96%D0%B2_%D1%96_%D0%B7%D0%B0%D1%85%D0%B8%D1%81%D0%BD%D0%B8%D1%86%D1%8C_%D0%A3%D0%BA%D1%80%D0%B0%D1%97%D0%BD%D0%B8
https://www.bbc.com/ukrainian/news-57836826

## 2022
In 2022, Day of Ukrainian Statehood was established.

Sources:
https://en.wikipedia.org/wiki/Statehood_Day_(Ukraine)
https://www.president.gov.ua/documents/4232021-39917

## 2023
In 2023, several holidays in Ukraine were changed:

- The Orthodox Christmas (_Різдво Христове_) on 7 January was abolished.
- Victory Day (Day of Remembrance and Victory over Nazism in World War II of 1939-1945, _День пам’яті та перемоги над нацизмом у Другій світовій війні 1939-1945 років_) was moved from 9 May to 8 May.
- Day of Ukrainian Statehood (_День Української Державності_) was moved from 28 June to 15 June.
- Day of Defenders of Ukraine (_День захисників і захисниць України_) was moved from 14 October to 1 October.

The law on Victory Day came into force on 15 June, 2023, so the change was effective from 8 May, 2024. The other three holidays were changed by an act which came into fore on 30 July, 2023, i.e. the Orthodox Christmas was last observed on 7 January, 2023, the adjusted Day of Ukrainian Statehood was first observed on 2024, and the adjusted Day of Defenders of Ukraine was first observed in 2023.

Sources:
https://zakon.rada.gov.ua/laws/show/3258-20
https://zakon.rada.gov.ua/laws/show/3107-20

Slava Ukraini! 🇺🇦