fix(ireland): New Year's Day on a Saturday also gives a substitute holiday (#375)
A “TODO” in the Ireland provider raises doubt whether a substitute holiday is added when New Year's Day falls on a Saturday.

The holiday was established in 1973 with a statutory instrument using the following wording: https://www.irishstatutebook.ie/eli/1974/si/341

> The first day of January if falling on a week day or, if not, the next day is hereby appointed to be a public holiday

I assume the ambiguity arises from whether Saturday is considered a weekday. However, the Holiday (Employees) Act, 1973, also uses the term “weekday”: https://www.irishstatutebook.ie/eli/1973/act/25/schedule/enacted/en/html

> (a) Christmas Day if falling on a weekday or, if not, the next Tuesday,
> (b) St. Stephen's Day if falling on a weekday or, if not, the next day,
> (c) St. Patrick's Day if falling on a weekday or, if not, the next day,

So I think it is fair to conclude that New Year's Day followed in same rules as the other holidays back then.

The new Organisation of Working Time Act, 1993 appear to have changed the rules on substitute holidays. AFAICT these are no longer moved to a specific Monday but may placed elsewhere at the discretion of the employer. However, it appears that most people have a day off on the same days as with the old rules.

https://www.irishstatutebook.ie/eli/1997/act/20/section/21/enacted/en/html

Examples:
https://www.ictu.ie/news/has-long-weekend-had-its-day-and-other-questions-about-bank-holidays-answered
https://bpfi.ie/wp-content/uploads/2022/01/Bank-Holidays-Republic-of-Ireland-2022-2024.pdf
https://constructionnews.ie/wp-content/uploads/2022/01/Builders-Holidays-2022.pdf
https://cravingcork.ie/public-holidays-ireland-2022/
https://www.fm104.ie/news/buzz/full-breakdown-bank-holidays-for-ireland-in-2022/

I guess you could argue that the substitute holidays should be marked with a different type than `TYPE_OFFICIAL`. I did not change that in this PR, though.

Signed-off-by: Christian Schmidt <github@chsc.dk>